+++ Magic/src/main/java/com/elmakers/mine/bukkit/magic/MageProperties.java
@@ -45,4 +45,8 @@ public class MageProperties extends CasterProperties {
     public Player getPlayer() {
         return mage.getPlayer();
     }
+
+    public Mage getMage() {
+        return mage;
+    }
 }