+++ Magic/CHANGELOG.md
@@ -9,6 +9,7 @@
  - Added support for SkillAPI attributes in spell configs
  - Added support for spells using SkillAPI mana (this is off by default, see config.yml defaults)
  - Fix setting properties on base classes (e.g. /mage configure mana_max 30)
+ - Fix mana buffs on armor (wizard armor)
 
 ## 7.0